requirements.txt
Django==4.2
Pillow
django-allauth==0.46.0
oauthlib==3.2.2
requests-oauthlib==1.3.1
django-environ==0.8.1
libscrc
qrcode